Fix pull request build issue
Travis runs two builds on pull request. The build due to commit does git 
push so the build due to pull request fials at git push as it needs to 
pull first. Just avoid git push on the pull request build as it adds no 
value anyway. The only reason to run it is to see if the build i.e. hugo 
command itself runs fine or not.